requirements.txt
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
botocore==1.38.27
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
makefun==1.16.0
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.5.0
pillow==11.2.1
propcache==0.3.2
pwdlib==0.2.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyvips==2.2.3
pyvips-binary==8.17.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
wrapt==1.17.2
yarl==1.20.1
